pyproject.toml
[project]
name = "pizzadelivery-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=6.0",
    "fastapi==0.111.0",
    "fastapi-jwt-auth==0.5.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary",
    "bcrypt==3.2.2",
    "pydantic>=2.7.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "sqlalchemy-orm>=1.2.10",
    "sqlalchemy-utils>=0.42.1",
    "sqlalchemy_utils",
    "utils>=1.0.2",
    "uvicorn>=0.30.0",
    "werkzeug>=3.1.5",
]
